tray icon refresh fellow style-name
