--- plugin-tray/effects.pri
@@ -1,8 +1,0 @@
-INCLUDEPATH += $$PWD
-INCLUDEPATH += $$PWD/..
-
-HEADERS += \
-    $$PWD/highlight-effect.h
-
-SOURCES += \
-    $$PWD/highlight-effect.cpp
